[OpenAPI] Rename "utils" -> "metatadata-getters"
