+++ packages/core/src/openapi/metadata-getters/get-api-callbacks.ts
