+++ packages/core/src/openapi/metadata-getters/get-api-complete-operation.ts
