+++ packages/core/src/openapi/metadata-getters/get-api-components.spec.ts
