+++ packages/core/src/openapi/metadata-getters/get-api-components.ts
