+++ packages/core/src/openapi/metadata-getters/get-api-deprecated.spec.ts
