+++ packages/core/src/openapi/metadata-getters/get-api-deprecated.ts
