+++ packages/core/src/openapi/metadata-getters/get-api-external-docs.spec.ts
