+++ packages/core/src/openapi/metadata-getters/get-api-operation.spec.ts
