+++ packages/core/src/openapi/metadata-getters/get-api-security.ts
