+++ packages/core/src/openapi/metadata-getters/get-api-used-tags.spec.ts
